refactor(mesh): modularise mesh/builder
